process_transcript/requirements.txt
google-cloud-storage==1.30.0
spacy==2.3.5
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz#egg=en_core_web_sm